refactor: move wheel handler to MainPanel per code review
Per Gemini's suggestion - move onWheel handler to MainPanel.tsx where
it can attach directly to scrollRef, eliminating the need for:
- MutationObserver retry logic
- document.querySelector('[data-main-content]')

Also extracts grid/zoom constants to shared utils/gridConstants.ts
for consistency across App, MainPanel, and ZoomSlider.
