+++ src/utils/gridConstants.ts
@@ -1,0 +1,11 @@
+// Grid zoom constants shared across App, MainPanel, and ZoomSlider
+export const GRID_SIZE_MIN = 80;
+export const GRID_SIZE_MAX = 320;
+export const GRID_SIZE_DEFAULT = 120;
+export const GRID_SIZE_STEP = 8;
+
+// Wheel event thresholds
+export const ZOOM_SCROLL_THRESHOLD = 50;
+export const ZOOM_BUFFER_RESET_MS = 200;
+export const NAV_SCROLL_THRESHOLD = 30;
+export const NAV_COOLDOWN_MS = 500;
